+++ src/main/java/org/swisspush/redisques/queue/QueueConsumerRunner.java
@@ -38,6 +38,7 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static java.lang.System.currentTimeMillis;
+import static org.swisspush.redisques.util.RedisquesAPI.BATCH_QUEUE;
 import static org.swisspush.redisques.util.RedisquesAPI.MESSAGE;
 import static org.swisspush.redisques.util.RedisquesAPI.OK;
 import static org.swisspush.redisques.util.RedisquesAPI.PAYLOAD;
@@ -78,7 +79,7 @@ public class QueueConsumerRunner {
         this.queueConfigurationProvider = queueConfigurationProvider;
 
         // handles trim request
-        trimRequestConsumer = vertx.eventBus().consumer(keyspaceHelper.getTrimRequestKey(), event -> {
+        trimRequestConsumer = vertx.eventBus().consumer(keyspaceHelper.getTrimRequestKey() + keyspaceHelper.getVerticleUid(), event -> {
             final String queueName = event.body();
             if (queueName == null) {
                 log.warn("Got event bus trim request msg with empty body! uid={}  address={}  replyAddress={}", keyspaceHelper.getVerticleUid(), event.address(), event.replyAddress());
@@ -238,7 +239,7 @@ public class QueueConsumerRunner {
         String queueKey = keyspaceHelper.getQueuesPrefix() + queueName;
         queueStatsService.createDequeueStatisticIfMissing(queueName);
         queueStatsService.dequeueStatisticSetLastDequeueAttemptTimestamp(queueName, System.currentTimeMillis());
-        processMessageWithTimeout(queueName, message, processResult -> {
+        processMessageWithTimeout(queueName, message, messageSize > 1, processResult -> {
 
             // update the queue failure count and get a retry interval
             int retryInterval = updateQueueFailureCountAndGetRetryInterval(queueName, processResult.getKey());
@@ -389,6 +390,7 @@ public class QueueConsumerRunner {
                                 if (delayed.failed()) {
                                     log.error("Delayed execution has failed.", exceptionFactory.newException(delayed.cause()));
                                 }
+                                setMyQueuesState(queueName, QueueState.READY);
                                 promise.complete();
                             });
                         }
@@ -477,7 +479,7 @@ public class QueueConsumerRunner {
         });
     }
 
-    private void processMessageWithTimeout(final String queue, final String payload, final Handler<Map.Entry<Boolean, String>> handler) {
+    private void processMessageWithTimeout(final String queue, final String payload, final boolean isBatch, final Handler<Map.Entry<Boolean, String>> handler) {
         long processorDelayMax = configurationProvider.configuration().getProcessorDelayMax();
         if (processorDelayMax > 0) {
             log.info("About to process message for queue {} with a maximum delay of {}ms", queue, processorDelayMax);
@@ -491,6 +493,9 @@ public class QueueConsumerRunner {
             String processorAddress = configurationProvider.configuration().getProcessorAddress();
             final EventBus eb = vertx.eventBus();
             JsonObject message = new JsonObject();
+            if (isBatch) {
+                message.put(BATCH_QUEUE, true);
+            }
             message.put("queue", queue);
             message.put(PAYLOAD, payload);
             log.trace("RedisQues process message: {} for queue: {} send it to processor: {}", message, queue, processorAddress);