+++ src/test/java/org/swisspush/redisques/action/GetPerQueueConfigurationsActionTest.java
@@ -38,7 +38,7 @@ public class GetPerQueueConfigurationsActionTest {
         Message<JsonObject> message = Mockito.mock(Message.class);
         ArgumentCaptor<JsonObject> captor = ArgumentCaptor.forClass(JsonObject.class);
         when(message.body()).thenReturn(new JsonObject("{\"operation\":\"getPerQueueConfiguration\",\"payload\":{\"filter\":\"*\"}}"));
-        QueueConfigurationProvider.provider(Vertx.vertx(), new ArrayList<>()).get().onComplete(event -> {
+        QueueConfigurationProvider.provider(Vertx.vertx(), new ArrayList<>(), 1_000).get().onComplete(event -> {
             QueueConfigurationProvider queueConfigurationProvider = event.result();
             GetPerQueueConfigurationsAction action = new GetPerQueueConfigurationsAction(queueConfigurationProvider, Mockito.mock(Logger.class));
 
@@ -64,7 +64,7 @@ public class GetPerQueueConfigurationsActionTest {
         Message<JsonObject> message = Mockito.mock(Message.class);
         ArgumentCaptor<JsonObject> captor = ArgumentCaptor.forClass(JsonObject.class);
         when(message.body()).thenReturn(new JsonObject("{\"operation\":\"getPerQueueConfiguration\",\"payload\":{\"configName\":\"test-pattern-3\"}}"));
-        QueueConfigurationProvider.provider(Vertx.vertx(), new ArrayList<>()).get().onComplete(event -> {
+        QueueConfigurationProvider.provider(Vertx.vertx(), new ArrayList<>(), 1_000).get().onComplete(event -> {
             QueueConfigurationProvider queueConfigurationProvider = event.result();
             GetPerQueueConfigurationsAction action = new GetPerQueueConfigurationsAction(queueConfigurationProvider, Mockito.mock(Logger.class));